create registration page

+++ base/views.py
@@ -1,6 +1,9 @@
 from django.shortcuts import render, redirect
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
+from django.http import HttpResponse
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
 from django.db.models import Q
@@ -17,8 +20,13 @@ from .forms import RoomForm
 
 def loginPage(request):
 
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -36,8 +44,22 @@ def loginPage(request):
         else:
             messages.error(request, 'Username and Password doesnot exits ')  
 
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_registration.html', context)
+
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An error occoure during re")
+    return render(request, 'base/login_registration.html', {'form': form})
 
 def logoutUser(request):
     logout(request)
@@ -65,7 +87,7 @@ def room(request, pk):
     #         room = i
     context = {'room': room}
     return render(request, 'base/room.html', context)
-
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -76,10 +98,13 @@ def create_room(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed !!!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -92,8 +117,14 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed !!!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
